Fix and improve PIO code
So far, the PIO code always both sampled the data bus _and_ drove the data bus - independently from the 65C02 R/W signal.

This means that during a 65C02 write cycle, both the 65C02 and the RP2040 drove the data bus - not. good. at. all.

Now the PIO code checks the 65C02 R/W signal and either samples the data bus _or_ drives the data bus.

A "side effect" is the improved performance as either the one or the other operation is performed.