Add rudimentary UART support
This likely wants to be made dependent on some preprocessor define or some configuration. But to decide on that is left to the project owner(s).

The UART is configured to the usual 115200,8,1,N on the UEXT connector pins designated for an UART.

The UART support can serve two purposes:

1. Console output and keyboard input for the 6502 kernel
2. Dev tracing - one can just use printf() and friends everywhere

Note: The initialisation in CONInitialise() was choosen because here
* The system clock is already set - which obligatory as otherwise the UART bit rate is not set correctly
* The prompt is not printed yet